python/requirements.txt
snowflake-connector-python>=3.6.0
pytest>=7.4.0
python-dotenv>=1.0.0

